LayoutTests/editing/smart-lists/backspace-after-typing-in-smart-list-does-not-undo-list-expected.txt

Once any other input follows the Smart Lists conversion, the immediate-undo window closes. Typing a character and then pressing Backspace should remove just that character; the list itself stays intact, with the second list item empty.
| <ul>
|   class="Apple-disc-list"
|   style="list-style-type: disc;"
|   <li>
|     "A"
|     <br>
|   <li>
|     <#selection-caret>
|     <br>
